FIX : [back] 엔 티 티 수 정

See merge request s08-webmobile1-sub2/S08P12E107!120

+++ backend/src/main/java/com/backend/db/entity/Badge.java
@@ -19,7 +19,7 @@ public class Badge {
 
     @Id
     @Column(name = "badge_sequence")
-    private Integer badgeSequence;
+    private Long badgeSequence;
     private String description;
     private String path;
